requirements.txt
echo "qrcode[pil]
pillow
streamlit"